requirements.txt
fastapi
uvicorn
web3
python-dotenv
pydantic
requests
streamlit
streamlit-autorefresh
pandas